\name{Wilting}
\docType{data}
\alias{Wilting}
\title{Cervical cancer arrayCGH data}
\description{
  A dataframe containing 4709 rows and 8 columns with arrayCGH data.
}
\usage{Wilting}
\format{
  A dataframe containing the following 8 columns:

  \describe{
    \item{Name}{The unique identifiers of array elements.}
    \item{Chromosome}{Chromosome number of each array element.}
    \item{Position}{Chromosomal position in bp of each array element.}
    \item{AdCA10}{Raw log2 ratios for cervical cancer sample AdCA10.}
    \item{SCC27}{Raw log2 ratios for cervical cancer sample SCC27.}
    \item{SCC32}{Raw log2 ratios for cervical cancer sample SCC32.}
    \item{SCC36}{Raw log2 ratios for cervical cancer sample SCC36.}
    \item{SCC39}{Raw log2 ratios for cervical cancer sample SCC39.}
  }
}
\source{
  Wilting, S.M., Snijders, P.J., Meijer, G.A., Ylstra, B., van den IJssel, P.R., Snijders, A.M., Albertson, D.G., Coffa, J., Schouten, J.P., van de Wiel, M.A., Meijer, C.J., & Steenbergen, R.D. (2006). Increased gene copy numbers at chromosome 20q are frequent in both squamous cell carcinomas and adenocarcinomas of the cervix. \emph{Journal of Pathology, 210}, 258-259.
}
\keyword{datasets}