\name{CGHcall-package}
\alias{CGHcall-package}
\docType{package}
\title{
	Calling aberrations for array CGH tumor profiles.
}
\description{
	
}
\details{
	\tabular{ll}{
	Package: \tab CGHcall\cr
	Type: \tab Package\cr
	Version: \tab 1.0\cr
	Date: \tab 2007-06-26\cr
	License: \tab GPL\cr
	}
}
\author{
Sjoerd Vosse & Mark van de Wiel

Maintainer: Sjoerd Vosse <s.vosse@vumc.nl>
}
\references{
  Mark A. van de Wiel, Kyung In Kim, Sjoerd J. Vosse, Wessel N. van Wieringen, Saskia M. Wilting and Bauke Ylstra. CGHcall: calling aberrations for array CGH tumor profiles. \emph{Bioinformatics, 23}, 892-894.
}
\keyword{ package }

